+++ AlquilerVehiculo/src/main/java/vehiculo/Furgoneta.java
@@ -3,7 +3,7 @@ package vehiculo;
 public class Furgoneta extends VehiculoDeCarga {
 
 	public Furgoneta(String matricula, String marca, Integer modelo, Integer pma) {
-		super(matricula, marca,modelo,pma, new Integer(500 + 300*pma));
+		super(matricula, marca,modelo,pma, new Double(500 + 300*pma));
 	}
 	
 }
