+++ AlquilerVehiculo/src/main/java/vehiculo/Microbus.java
@@ -3,11 +3,11 @@ package vehiculo;
 public class Microbus extends VehiculoDePasageros{
 
 	public Microbus(String matricula, String marca, Integer modelo, Integer cantidadPlazas) {
-		super(matricula, marca, modelo, cantidadPlazas, new Integer(1500));
+		super(matricula, marca, modelo, cantidadPlazas, new Double(1500));
 		
 	}
 	
-	public Integer obtenerPrecio(Integer cantidadDias) {
+	public Double obtenerPrecio(Integer cantidadDias) {
 		
 		return (this.componenteBasePorDia + this.plazas*120 ) * cantidadDias;
 	}
