+++ AlquilerVehiculo/src/main/java/vehiculo/VehiculoDeCarga.java
@@ -4,7 +4,7 @@ public abstract class VehiculoDeCarga extends Vehiculo{
 	
 	protected Integer pma;
 	
-	public VehiculoDeCarga(String matricula, String marca, Integer modelo, Integer pma, Integer componenteBasePorDia) {
+	public VehiculoDeCarga(String matricula, String marca, Integer modelo, Integer pma, Double componenteBasePorDia) {
 		super(matricula, marca, modelo, componenteBasePorDia);
 		this.pma = pma ;
 		
@@ -15,7 +15,7 @@ public abstract class VehiculoDeCarga extends Vehiculo{
 		return this.pma;
 	}
 	
-	public Integer obtenerPrecio(Integer cantidadDiasAlquiler) {
+	public Double obtenerPrecio(Integer cantidadDiasAlquiler) {
 		
 		return (( this.componenteBasePorDia)*cantidadDiasAlquiler); 
 	}
